move integration tests into server project to make sure server binary is built before they run
